refactored Header
